Refactor the search dialog form to be a component
